pyproject.toml
[build-system]
requires = ["setuptools>=67", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "factbase-tool"
version = "0.1.0"
description = "Roll Call Factba.se transcripts scraper, indexer, and web UI"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "Factbase Tool"}]
dependencies = [
  "Flask==2.3.3",
  "Typer==0.9.0",
  "httpx==0.27.0",
  "beautifulsoup4==4.12.3",
  "lxml==5.2.1",
  "python-dateutil==2.9.0.post0",
  "playwright==1.42.0",
  "rich==13.7.0",
  "uvloop==0.19.0; platform_system != 'Windows'",
]

[project.scripts]
factbase = "factbase.cli:app"

[tool.setuptools]
packages = ["factbase"]

